+++ src/components/formStore/FormStoreTableProvider.tsx
@@ -21,6 +21,15 @@ import NoResourcesYet from '../messages/NoResourcesYet'
 import FormStoreIcon from './display/FormStoreIcon'
 import { FormStoreTableContext } from './useFormStoreTableContext'
 
+const TABLE_PARAMETERS_FORM_STORE_KEY = 'REACT_TABLE_PARAMETERS_FORM_STORE'
+
+function getParamsFromLocalStorage() {
+  const text = localStorage.getItem(TABLE_PARAMETERS_FORM_STORE_KEY)
+  if (text) {
+    return JSON.parse(text)
+  }
+}
+
 export function FormStoreTableProvider({
   form,
   children,
@@ -48,6 +57,11 @@ export function FormStoreTableProvider({
     debounceSearchMs: 1000,
     onDefaultFilters: React.useCallback((query) => {
       let defaultParameters: formStoreService.FormStoreParameters = {}
+      const localStorageParams = getParamsFromLocalStorage()
+      if (localStorageParams?.unwindRepeatableSets === true) {
+        defaultParameters.unwindRepeatableSets =
+          localStorageParams.unwindRepeatableSets
+      }
       try {
         if (typeof query.parameters === 'string') {
           defaultParameters = JSON.parse(query.parameters)
@@ -108,6 +122,16 @@ export function FormStoreTableProvider({
   const submissionIdValidationMessage = useSubmissionIdValidationMessage(
     parameters.filters?.submissionId?.$eq,
   )
+
+  React.useEffect(() => {
+    const paramsToStore = {
+      unwindRepeatableSets: parameters?.unwindRepeatableSets,
+    }
+    localStorage.setItem(
+      TABLE_PARAMETERS_FORM_STORE_KEY,
+      JSON.stringify(paramsToStore),
+    )
+  }, [parameters])
 
   const formStoreTable = useFormStoreTable({
     formStoreRecords,