+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - **[BREAKING]** `allowSubmitWithPendingAttachments` prop for `<OneBlinkForm />` component to `isPendingQueueEnabled` to prevent offline submissions. The functionality regarding attachments from the previous `allowSubmitWithPendingAttachments` prop have not changed.
 - `FormNotification` to handle static data lookup type
+- `unwindRepeatableSets` parameter stored in localStorage
 
 ## [3.1.0] - 2023-06-05